+++ bricks/response_dto/CHANGELOG.md
@@ -1,3 +1,10 @@
-# 0.1.0+1
+# 0.0.1
 
-- TODO: Describe initial release.
+- Initial commit
+
+# 0.0.2
+
+- Added empty factory
+- Made all variables nullable
+- Updated the fromJson method with type condition
+- Updated README